Moved utility functions to their own submodule to prevent circular initiation.
