+++ satkit/metrics/aggregate_image.py
@@ -40,7 +40,7 @@ import numpy as np
 from satkit.data_structures import (
     FileInformation, Modality, Recording, Session, Statistic, StatisticMetaData
 )
-from satkit.helpers import product_dict
+from satkit.utility_functions import product_dict
 
 _logger = logging.getLogger('satkit.aggregate_image')
