+++ satkit/utility_functions/processing_helpers.py
