+++ satkit/utility_functions/types.py
